+++ CHANGES.rst
@@ -10,7 +10,7 @@ astropy.config
 astropy.constants
 ^^^^^^^^^^^^^^^^^
 
-- Add IERS2010 numerical standarts (constants). [#6715]
+- Add IERS2010 numerical standards (constants). [#6715]
 
 astropy.convolution
 ^^^^^^^^^^^^^^^^^^^